refactor airo-dataset-tools package to prepare for labeling code
