+++ airo-dataset-tools/test/test_coco_load.py
@@ -13,7 +13,7 @@ import json
 import os
 
 import pytest
-from airo_dataset_tools.coco.coco_parser import (
+from airo_dataset_tools.data_parsers.coco import (
     CocoCategory,
     CocoInstancesDataset,
     CocoKeypointCategory,
